requirements.txt
flask
pandas
python-dotenv
langchain
langchain-openai
gunicorn
